+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeAttributeValueType.kt
@@ -23,7 +23,6 @@ enum class LimeAttributeValueType(private val tag: String) {
     NAME("Name"),
     ACCESSORS("Accessors"),
     ATTRIBUTE("Attribute"),
-    BUILDER("Builder"),
     CONST("Const"),
     CSTRING("CString"),
     DEFAULT("Default"),